chore: rename language option
It makes more sense to use JavaScript instead of Node.js as the label
for the language. While initially this was meant for Electron, when
working on React Native I realized it made more sense to change it and
maybe add a note to clarify what runtime is used (Node.js, Hermes, etc.)
where applicable.