chore: reflect changes in technology data faster
Execute the technology update GitHub Action every 15 minutes.
Clean the build folder when building in Netlify. For some reason even if
the changes in the JSON are commited and picked up by Netlify, the
website still uses the previous version of the data. I suspect that
Docusaurus does not see any change in the markdown and thus uses the
same `build` folder as production. This is similar to what happens with
its watch system when running `npm start`: There has to be a change in a
md file to be picked up and reflected.